requirements.txt
Django==1.9.5
django-crispy-forms==1.7.0
django-filter==1.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.4
djangorestframework==3.3.1
Markdown==2.6.11
markdown2==2.3.5
Pillow==5.0.0
gunicorn==19.8.1
psycopg2==2.7.4
redis==2.10.5   
django-guardian
django-cors-headers==2.2.0
eventlet
